Added a functional AI.
Added the AI component, which uses minimax with alpha-beta pruning to select moves.  Also added the mechanism for the AI moving after the player moves, and set up functionality for either
player starting or computer starting, though this is currently selectable only in code; UI functionality will be created next.